Add a force flag (-f) to the run.sh script
That means rebuild the container even if the image is current.

+++ system/run.sh
@@ -1,14 +1,33 @@
 #!/usr/bin/env bash
 
 # Start the mirror server the first time or update to a newer version.
+# If run with -f, force the update and restart even if the image is up to date.
 
 set -euo pipefail
+
+# Check for the force flag
+force=false
+while getopts "f" opt; do
+    case ${opt} in
+        f)
+            force=true
+            ;;
+        \?)
+            echo "Usage: $0 [-f]"
+            exit 1
+            ;;
+    esac
+done
+
 
 echo "> Pulling latest image"
 if docker pull "genericmoniker/mirror:main" | grep "Image is up to date" > /dev/null; then
     echo "> Image is up to date"
-    echo "> Done"
-    exit 0
+    # If the image is up to date and we're not forcing an update, exit.
+    if [ "$force" = false ]; then
+        echo "> Done"
+        exit 0
+    fi
 fi
 
 echo "> Removing previous container"
